feat: [HOMETASK-1] implement quote web server
- QuoteService implements/extends repo
- JsonDb as an independent module
- style
