+++ apps/api/src/app/quotes/services/index.ts
@@ -1,2 +1,1 @@
-export * from './json-db-collection/json-db-collection.service';
 export * from './quotes.service';
